refactor: logger interfaces
