+++ CHANGELOG.md
@@ -6,6 +6,13 @@ All notable changes to this project will be documented in this file.
 
 ### Breaking changes
 
+* To work with `log` crate, now you need to set both:
+    ```rust
+    fn main() {
+        logforth::bridge::setup_log_crate();
+        logforth::builder().apply()
+    }
+    ```
 * All interfaces that return `anyhow::Result` is now using a result over `logforth::Error`.
 * `JsonLayout` now collects diagnostics context into a separate field `diags`.
 * `SingleFile` appender is removed. You can replace it with `append::File`.
@@ -39,7 +46,6 @@ All notable changes to this project will be documented in this file.
 * `EnvFilter` is now self-hosted. Some methods may be changed, but the general user experience should retain:
   * `EnvFilter`'s constructors (`from_env`, etc.) are moved to `EnvFilterBuilder`.
 * Upgrade to opentelemetry 0.31.0.
-* `LoggerBuilder`'s `try_apply`/`apply` methods are renamed to `try_setup_log_crate`/`setup_log_crate`.
 * `TextLayout` is now behind `layout-text` feature flag, and colored is always available when the feature is enabled.
 * Internal log structs are migrated from `log` crate to self-hosted types. This should not affect most users, but if you are customizing appender, layout, filter, and diagnostic, you should replace `log::Record`, `log::Metadata`, or `log::Level`, with `logforth::Record`, `logforth::Metadata`, or `logforth::Level`.
